updated CLM model and test is added

+++ config/lcquad_config.py
@@ -31,8 +31,9 @@ class LCQuadConfig:
         model_config = self.load_yaml_with_jinja("config/lcquad_model_config.yaml")
 
         model_config['model']['device'] = torch.device("mps" if torch.mps.is_available() else "cpu")
+        # model_config['model']['device'] = "cpu" # using cpu to use F32
         # for LoRA - SFT, only fp16 is supported by MPS system
-        model_config['model']['dtype'] = torch.float16
+        model_config['model']['dtype'] = torch.float32
 
         model_ind = model_config['model']['chosen_model']
         model_version = "latest"
